Install packages - update for 1 record
Added function to check for packages that are installed and install if
they are required for extension.

Added if statement so if only one location is being geocoded there is a
check and the results will be put into a table.  Previously the function
broke if there was only one location being searched for.